% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltools.R
\name{absmin}
\alias{absmin}
\title{Return the absolute minimum of a numeric vector}
\usage{
absmin(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The absolute minimum value of \code{x}.
}
\description{
This function returns the absolute minimum of a numeric vector.
}
\author{
Wolfram Hoeps
}
