% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_bed_format_and_qc.R
\name{add_and_fix_columns}
\alias{add_and_fix_columns}
\title{A helperfunction (2/2) for the process of turning a bed to tsv.}
\usage{
add_and_fix_columns(sds)
}
\arguments{
\item{sds}{\link{data frame} an intermediate sds file between bed and tsv. Produced by
make_sd_partners, the other helperfunction.}
}
\value{
a tsv dataframe
}
\description{
A helperfunction (2/2) for the process of turning a bed to tsv.
}
\author{
Wolfram Höps
}
