% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{add_annotation_blocks}
\alias{add_annotation_blocks}
\title{Add annotation blocks to a given ggplot object}
\usage{
add_annotation_blocks(gp, opt)
}
\arguments{
\item{gp}{A ggplot object to add annotation blocks to}

\item{opt}{A list containing necessary options for the annotation}
}
\value{
A ggplot object with annotation blocks added
}
\description{
Add annotation blocks to a given ggplot object
}
