% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{add_eval}
\alias{add_eval}
\title{add_eval
CRITICAL: NEEDS IMPROVEMENTS}
\usage{
add_eval(res, m, layer, pair_level1, pair_level2, pair_level3)
}
\arguments{
\item{bitlocus}{matrix, nxm}

\item{depth}{How many SVs in sequence should be simulated?}
}
\value{
evaluation matrix
}
\description{
A helperfunction, wrapping the evaluation function.
}
\author{
Wolfram Höps
}
