% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{add_file_highlight}
\alias{add_file_highlight}
\title{Add file-based highlights to a ggplot object}
\usage{
add_file_highlight(gp, opt)
}
\arguments{
\item{gp}{A ggplot object to add file-based highlights to}

\item{opt}{A list containing necessary options for the highlighting}
}
\value{
A ggplot object with file-based highlights added
}
\description{
Add file-based highlights to a ggplot object
}
