% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder.R
\name{add_snps}
\alias{add_snps}
\title{Add snps to a simulated SD pair}
\usage{
add_snps(seq, similarity, seed = 1234)
}
\arguments{
\item{seq}{DNA sequence, as 'character'}

\item{similarity}{How similar should the result be to the input (fraction of bases, \link{0-1})}
}
\description{
Mutate a proportion of bases in a DNA sequence. Can be used to
simulate evolutionary divergence of two SD pairs.
}
\author{
Wolfram Höps
}
