% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_whole_genome_list.R
\name{align_all_vs_all_using_minimap2}
\alias{align_all_vs_all_using_minimap2}
\title{align_all_vs_all_using_minimap2}
\usage{
align_all_vs_all_using_minimap2(
  minimap2_bin,
  reference_fasta,
  asm_fasta,
  out_paf,
  threads
)
}
\arguments{
\item{minimap2_bin}{Path to the minimap2 binary}

\item{reference_fasta}{Path to the reference fasta}

\item{asm_fasta}{Path to the assembly fasta}

\item{out_paf}{Path to the output paf file}

\item{threads}{Number of threads to use}
}
\value{
Nothing
}
\description{
align_all_vs_all_using_minimap2
}
