% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_whole_genome_list.R
\name{align_all_vs_all_using_minimap2_shred_merge}
\alias{align_all_vs_all_using_minimap2_shred_merge}
\title{align_all_vs_all_using_minimap2}
\usage{
align_all_vs_all_using_minimap2_shred_merge(
  awkscript,
  reference_fasta,
  asm_fasta,
  out_paf,
  threads,
  outdir
)
}
\arguments{
\item{reference_fasta}{Path to the reference fasta}

\item{asm_fasta}{Path to the assembly fasta}

\item{out_paf}{Path to the output paf file}

\item{threads}{Number of threads to use}

\item{minimap2_bin}{Path to the minimap2 binary}
}
\value{
Nothing
}
\description{
align_all_vs_all_using_minimap2
}
