% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{annotate_pairs_with_hash}
\alias{annotate_pairs_with_hash}
\title{annotate_pairs_with_hash}
\usage{
annotate_pairs_with_hash(bitlocus, pairs)
}
\arguments{
\item{bitlocus}{input bitlocus; a matrix}

\item{pairs}{a data frame (?) containing known hashes.}
}
\description{
To speed up grid search,, we keep an inventory of bitloci which we have seen.
For this, we simplify the matrix ('return diag values') and then turn this in to a hash,
which we save and can use for comparisons later.
}
