% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{annotate_res_out_with_positions_lens}
\alias{annotate_res_out_with_positions_lens}
\title{annotate_res_out_with_positions_lens}
\usage{
annotate_res_out_with_positions_lens(res_out, bitlocus)
}
\description{
A helper function that takes a results dataframe and an original bitlocus,
and translates the results (e.g. '4-10-inv') into coordinates (e.g. 450000 - 600000 - inv).
This is quite limited and rudimentary at this point.
}
