% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{areColors}
\alias{areColors}
\title{Check if given character vector elements are valid color representations}
\usage{
areColors(x)
}
\arguments{
\item{x}{A character vector of color names, hex codes or RGB values.}
}
\value{
A logical vector of the same length as \code{x}, indicating whether each element is a valid color representation.
}
\description{
This function checks if the given input is a valid color name, hex code or RGB value.
}
\details{
The function replaces missing values with 'none' and tries to convert each element to an RGB matrix using \code{col2rgb}.
If successful, the element is considered a valid color representation.
}
\examples{
areColors(c("red", "#FF0000", "rgb(255, 0, 0)", "blue", "invalid"))
areColors(c("green", "#00FF00", "rgb(0, 255, 0)", NA))

}
