% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{awk_edit_paf}
\alias{awk_edit_paf}
\title{Submit a system command to run awk to change sequence chunk names}
\usage{
awk_edit_paf(inpaf, outpaf)
}
\arguments{
\item{inpaf}{\link{character/link} input paf}

\item{outpaf}{\link{character/link} output paf}
}
\value{
nothing. Only output files written.
}
\description{
This is a helperfunction to run awk
}
\author{
Wolfram Höps
}
