% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_vis.R
\name{benchmark_vis}
\alias{benchmark_vis}
\title{Run a benchmark visualization. Thrown into a function but is really a script.}
\usage{
benchmark_vis()
}
\description{
TO BE REVIEWED / CORRECTED!
}
\author{
Wolfram Höps
}
