% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{bounce_point}
\alias{bounce_point}
\title{bounce point. Algorithm for making the grid.
Description TBD.}
\usage{
bounce_point(vectors, point)
}
\description{
bounce point. Algorithm for making the grid.
Description TBD.
}
\author{
Wolfram Hoeps
}
