% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltools.R
\name{calc_coarse_grained_aln_score}
\alias{calc_coarse_grained_aln_score}
\title{calc_coarse_aln_score}
\usage{
calc_coarse_grained_aln_score(
  mat,
  old_way_of_calc = F,
  verbose = F,
  forcecalc = F,
  orig_symm = 1
)
}
\arguments{
\item{mat}{matrix. Bitlocus / Gridmatrix.}
}
\description{
A function assigning an alignment score to a condensed dotplot.
we are using a base-function from geeksforgeeks.org, which is
essentially like needleman-wunsch.
Since in our case the costs for traversing a node 'vertically' or
'horizontally' are different, we had to work around this a bit by
adapting the algorithm.
}
\author{
geeksforgeeks.org; Wolfram Hoeps
}
