% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{calc_symm}
\alias{calc_symm}
\title{Calculate the initial symmetry of a bitlocus}
\usage{
calc_symm(bitl)
}
\arguments{
\item{bitl}{A data frame representing a bitlocus}
}
\value{
The initial symmetry of the bitlocus
}
\description{
Calculate the initial symmetry of a bitlocus
}
