% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltools.R
\name{calculate_estimated_aln_score}
\alias{calculate_estimated_aln_score}
\title{calculate_estimated_aln_score}
\usage{
calculate_estimated_aln_score(mat)
}
\arguments{
\item{mat}{A matrix object representing a condensed dotplot}
}
\description{
Quick & Dirty estimation of the score of a condensed dotplot.
Is used to filter out hopeless cases.
}
\author{
Wolfram Hoeps
}
