% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{carry_out_compressed_sv}
\alias{carry_out_compressed_sv}
\title{Find SV carry_out_compressed_sv}
\usage{
carry_out_compressed_sv(bitl, input_ins)
}
\arguments{
\item{bitl}{an nxm matrix (n=gridpoints_x, m=gridpoints_y)}

\item{input_ins}{a vector with instruction: p1, p2, sv.}
}
\description{
TBD
}
\author{
Wolfram Höps
}
