% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{carry_out_del}
\alias{carry_out_del}
\title{Introduce a DEL to a DNA sequence}
\usage{
carry_out_del(seq, sds, sv)
}
\arguments{
\item{seq}{\link{character} DNA sequence as character vector}

\item{sds}{\link{data.frame} A dataframe with tsv SD information}

\item{sv}{\link{data.frame} A data framewith columns \link{SD-uid, SV}.
SD-uid has to be existing in sds_tsv. SV can be \link{INV, DEL, DUP}}
}
\value{
a list object containing the mutated sequence and sds dataframe.
}
\description{
This function is typically called by the mutate_seq function
}
\author{
Wolfram Höps
}
