% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{clean_after_yourself}
\alias{clean_after_yourself}
\title{clean_after_yourself}
\usage{
clean_after_yourself(outlinks)
}
\description{
This function, if invoked, deletes intermediate fasta sequences created by NAHRwhals.
By default, this is not invoked.
}
