% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{colMax}
\alias{colMax}
\title{colMax function}
\usage{
colMax(data)
}
\arguments{
\item{data}{a matrix or data frame.}
}
\value{
a numeric vector containing the maximum value for each column of \code{data}.
}
\description{
Computes the maximum value for each column of a matrix or data frame.
}
