% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_paf.R
\name{compress_paf_fnct}
\alias{compress_paf_fnct}
\title{Melt the alignment pieces back together after chunkifying them earlier.}
\usage{
compress_paf_fnct(
  inpaf_link = NULL,
  outpaf_link = NULL,
  inpaf_df = NULL,
  save_outpaf = T,
  n_quadrants_per_axis = NULL,
  second_run = F,
  inparam_chunklen = NULL,
  inparam_compression = NULL
)
}
\arguments{
\item{inpaf_link}{\link{character/link} link to the chunked paf}

\item{outpaf_link}{\link{character/link} link to the molten output paf
Set to larger values (e.g. 250k) for very large alignments (>5Mb).}
}
\description{
This is a crucial module for finding the SDs later, and poses
the last part of the chunk-minimap2-edit-melt workflow. This might need more
work in the future, because melting is not so straight forward.
}
\author{
Wolfram Höps
}
