% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{confirm_loaded_nahr}
\alias{confirm_loaded_nahr}
\title{A testfunction to see if nahrtoolkit is loaded}
\usage{
confirm_loaded_nahr()
}
\description{
A testfunction to see if nahrtoolkit is loaded
}
\author{
Wolfram Höps
}
