% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{confirm_loaded_nw}
\alias{confirm_loaded_nw}
\title{A testfunction to see if nahrwhals is loaded}
\usage{
confirm_loaded_nw()
}
\description{
A testfunction to see if nahrwhals is loaded
}
\author{
Wolfram Hoeps
}
