% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{correct_paf}
\alias{correct_paf}
\title{Corrects a PAF file if it's made on scattered input.}
\usage{
correct_paf(inputpaf, outputpaf)
}
\arguments{
\item{inputpaf}{A character string specifying the path to the input PAF file.}

\item{outputpaf}{A character string specifying the path to the output PAF file.}
}
\value{
Nothing is returned by the function, but the corrected PAF file is saved to the output file.
}
\description{
This function takes an input PAF file and an output PAF file as arguments,
and runs an 'awk' command to correct the input PAF file if it's made on
scattered input. The corrected PAF file is saved to the output file.
}
\author{
Wolfram Hoeps
}
