% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{create_vcf_line}
\alias{create_vcf_line}
\title{Create each line of the VCF}
\usage{
create_vcf_line(df, GTS)
}
\arguments{
\item{df}{A data frame from the generate_gt_strings function}

\item{GTS}{A data frame with the genotype strings}
}
\value{
A data frame with a single VCF line
}
\description{
Create each line of the VCF with the genotype string
}
\author{
Wolfram Hoeps
}
