% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{decide_loop_continue_symmetry}
\alias{decide_loop_continue_symmetry}
\title{decide_loop_continue_symmetry}
\usage{
decide_loop_continue_symmetry(bitl_f, symm_cutoff = 0.8, orig_symm = 1)
}
\arguments{
\item{bitl_f}{a condensed bitlocus as a matrix}

\item{symm_cutoff}{a value between 0 and 1 for lowest acceptable symmetry (default: 0.8)}

\item{orig_symm}{remnant from debugging, don't touch please :)}
}
\description{
In the tree search, we sometimes want to stop exploring an outcome if the dotplot is very unsymmetrical.
Symmetricity (is that a word?) is one of the simplest proxy for deciding if a dotplot is high or low score.
}
