% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{define_output_files}
\alias{define_output_files}
\title{Define output file paths for NAHRwhals output}
\usage{
define_output_files(sequence_name_output, samplename)
}
\arguments{
\item{sequence_name_output}{The output directory path for the files.}

\item{samplename}{The name of the sample being analyzed.}
}
\value{
A list of file paths for the output files.
}
\description{
This function defines output file paths for various intermediate and final files generated by NAHRwhals.
}
