% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_machinery.R
\name{dfs}
\alias{dfs}
\title{DFS caller function
This is the gateway into entering the dfs.}
\usage{
dfs(bitlocus, maxdepth = 3, increase_only = F, earlystop = Inf, history = NULL)
}
\description{
DFS caller function
This is the gateway into entering the dfs.
}
