% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{diagonalize}
\alias{diagonalize}
\title{Diagonalize}
\usage{
diagonalize(matrix, factor = 0.25)
}
\arguments{
\item{matrix}{a matrix typically representing a bitlocus}

\item{factor}{value between 0 and 1 determining how much to erase from the matrix.}
}
\description{
Takes a matrix, turns the corners into zero.
}
