% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafCoordsDotPlotly.R
\name{dotplotly_dotplot}
\alias{dotplotly_dotplot}
\title{A core function for plotting a paf.}
\usage{
dotplotly_dotplot(opt)
}
\arguments{
\item{opt}{a list of parameters.}
}
\value{
Nothing, but writes a pdf with a dotplot.
}
\description{
This is extracted and an edited version of
the script "pafCoordsDotPlotly.R" from the dotplotly package
(https://github.com/tpoorten/dotPlotly) by Tom Poorten.
Should be run through the 'pafdotplot_make' wrapper function
which helps in selecting input parameters.
}
\author{
Tom Poorten, edited by Wolfram Höps
}
