% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{dotplotly_dotplot_return_aln}
\alias{dotplotly_dotplot_return_aln}
\title{A core function for plotting a paf.}
\usage{
dotplotly_dotplot_return_aln(opt)
}
\arguments{
\item{opt}{a list of parameters.}
}
\value{
Nothing, but writes a pdf with a dotplot.
}
\description{
This is extracted and an edited version of
the script "pafCoordsDotPlotly.R" from the dotplotly package
(https://github.com/tpoorten/dotPlotly) by Tom Poorten.
Should be run through the 'pafdotplot_make' wrapper function
which helps in selecting input parameters.
}
\author{
Tom Poorten (https://github.com/tpoorten/dotPlotly), edited by Wolfram Höps.
}
