% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqplotter.R
\name{dotPlotr}
\alias{dotPlotr}
\title{A helperfunction for making an exact dotplot, wrapping a functionality of dotplot package.
Should not be used for sequences above 25 kbp.}
\usage{
dotPlotr(seq1, seq2, wsize = 5, wstep = 1, nmatch = -1)
}
\arguments{
\item{seq1}{A DNA sequence, as character or DNAString}

\item{seq2}{A DNA sequence, as character or DNAString}
}
\value{
a ggplot2 object: exact dotplot between the two seqs.
}
\description{
Give me two sequences (character of DNAString I think),
and I will return to you a ggplot object with an exact (bp-precise)
dotplot. This relies on the mkDotPlotDataFrame function from 'dotplot' package.
}
\author{
Wolfram Höps
}
