% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{eval_mutated_seq}
\alias{eval_mutated_seq}
\title{eval_mutated_seq}
\usage{
eval_mutated_seq(bitlocus)
}
\arguments{
\item{bitlocus}{an nxm matrix (n=gridpoints_x, m=gridpoints_y)}
}
\value{
numeric \link{0-1} with evaluation score.
}
\description{
Evaluation function: how similar are the input and output?
This needs work still...
}
\author{
Wolfram Höps
}
