% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_vis.R, R/sd_compare_input_output.R
\name{benchmark_vis}
\alias{benchmark_vis}
\alias{turn_intervals_to_vector_0_1}
\alias{compare_sd_paf_input_output}
\title{Run a benchmark visualization. Thrown into a function but is really a script.}
\usage{
benchmark_vis()

turn_intervals_to_vector_0_1(intervals, maxlen)

compare_sd_paf_input_output(
  preslink,
  ptruelink,
  chunklen,
  runname,
  strand,
  outfile
)
}
\arguments{
\item{intervals}{\link{data.frame} A data frame with interval columns \link{start, stop}}

\item{maxlen}{\link{character/link/NULL} Total length of the vector to create. Should
be >= max(invervals$stop)}

\item{preslink}{\link{character/link} a link to a 'results' paf (typically minimap2 output
or processed output)}

\item{chunklen}{\link{numeric} chunk length. This is for naming purposes only. Not so greatly implemented.}

\item{runname}{\link{character} name}

\item{strand}{\link{"+"/"-"} will also appear in output file}

\item{outfile}{the results .txt file which will be written to.}

\item{ptruelink[character/link]}{a link to a ground truth tsv file (single or double entries though?)}
}
\value{
a vector of length maxlen, with 0/1 depending on coverage of intervals.

nope
}
\description{
TO BE REVIEWED / CORRECTED!

takes a list of intervals, and a max sequence length. Returns
a 0-1 vector where 1 is spaces covered by an SD. This is a quite inefficient
algorithm, but the part is not extremely time-sensitive.


}
\author{
Wolfram Höps

Wolfram Höps

Wolfram Höps
}
