% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{expand_one_of_pair}
\alias{expand_one_of_pair}
\title{Helperfunction 1/2 for carry_out_dup}
\usage{
expand_one_of_pair(sds, uid, n_of_pair, dupsize)
}
\description{
Helperfunction 1/2 for carry_out_dup
}
\author{
Wolfram Höps
}
