% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{expand_whole_pair}
\alias{expand_whole_pair}
\title{Helperfunction 2/2 for carry_out_dup}
\usage{
expand_whole_pair(sds, uid, dupsize)
}
\description{
Helperfunction 2/2 for carry_out_dup
}
\author{
Wolfram Höps
}
