% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{explore_mutation_space}
\alias{explore_mutation_space}
\title{explore_mutation_space}
\usage{
explore_mutation_space(bitlocus, depth)
}
\arguments{
\item{bitlocus}{matrix, nxm}

\item{depth}{How many SVs in sequence should be simulated?}
}
\value{
evaluation matrix
}
\description{
Main workhorse, tying together the pieces.
}
\author{
Wolfram Höps
}
