% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seq_wrapper.R
\name{extract_sequence_wrapper}
\alias{extract_sequence_wrapper}
\title{A wrapper for extracting sequence}
\usage{
extract_sequence_wrapper(params, outlinks)
}
\description{
Needs more documentation.
}
\author{
Wolfram Höps
}
