% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{extract_subseq}
\alias{extract_subseq}
\title{extract_subseq
helperfunction to extract a subfasta by coordinates}
\usage{
extract_subseq(infasta, seqname, start, end, outfasta)
}
\description{
Is awfully slow unfortunately. We should find a way to only read
the chromosome of interest.
}
\author{
Wolfram Hoeps
}
