% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{find_coords_mad}
\alias{find_coords_mad}
\title{find_coords_mad}
\usage{
find_coords_mad(liftover_coords, cpaf, winner_chr, start, end, liftover_run)
}
\description{
find_coords_mad
}
\author{
Wolfram Hoeps
}
