% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{find_maxdiag}
\alias{find_maxdiag}
\title{find_maxdiag}
\usage{
find_maxdiag(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
Scan a matrix for the length of the longest (positive) diagonal. This is part
of the 'eval' method.
}
\author{
Unknown (https://stackoverflow.com/questions/47330812/find-the-longest-diagonal-of-an-element-in-a-matrix-python)
}
