% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_find_parameters.R
\name{find_minlen_compression_params_wiggle}
\alias{find_minlen_compression_params_wiggle}
\title{Find parameters by wiggling.
This produces unstable and unpredictable results; use is discouraged but
we are leaving the function alive for now. Is called when compression_params$auto_find is set to True (default: False)}
\usage{
find_minlen_compression_params_wiggle(inpaf, compression_params)
}
\description{
Find parameters by wiggling.
This produces unstable and unpredictable results; use is discouraged but
we are leaving the function alive for now. Is called when compression_params$auto_find is set to True (default: False)
}
