% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{find_sv_opportunities}
\alias{find_sv_opportunities}
\title{Find SV opportunities}
\usage{
find_sv_opportunities(sample)
}
\arguments{
\item{sample}{an nxm matrix (n=gridpoints_x, m=gridpoints_y)}
}
\value{
A dataframe with columns 'p1', 'p2', 'sv'
}
\description{
Building block: Take in a 'bitlocus',
and return a list of NAHR-driven rearrangements that
are possible on the reference sequence.
}
\author{
Wolfram Höps
}
