% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{find_threshold}
\alias{find_threshold}
\title{find_threshold}
\usage{
find_threshold(bitlocus_f, nmut)
}
\arguments{
\item{bitlocus_f}{The matrix representing the bitlocus.}

\item{nmut}{The number of mutations.}
}
\value{
The threshold.
}
\description{
This function calculates the alignment threshold for determining when a bitlocus is considered "solved".
}
