% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_bitl_y_if_needed.R
\name{flip_bitl_y_if_needed}
\alias{flip_bitl_y_if_needed}
\title{flip_bitl_y_if_needed}
\usage{
flip_bitl_y_if_needed(bitl)
}
\arguments{
\item{bitl}{a gridmatrix / bitlocus.}
}
\value{
the same gridmatrix, flipped or unflipped.
}
\description{
Determines whether a bitlocus is 'upside down', which happens when the
alignment is negative. To determine this, look only at values on the outer
25\% edge of the plot (assuming that SVs are mainly confined to the middle),
and determine if this outer part has more positive or negative alignments.
If there are more negative, then we flip the y-axis.
}
\author{
Wolfram Hoeps
}
