% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_bed_format_and_qc.R
\name{add_and_fix_columns}
\alias{add_and_fix_columns}
\alias{make_sd_partners}
\alias{zipFastener}
\alias{run_qc_bed_df}
\title{A helperfunction (2/2) for the process of turning a bed to tsv.}
\usage{
add_and_fix_columns(sds)

make_sd_partners(sds_raw)

zipFastener(df1, df2, along = 2)

run_qc_bed_df(sds_raw)
}
\arguments{
\item{sds}{\link{data frame} an intermediate sds file between bed and tsv. Produced by
make_sd_partners, the other helperfunction.}

\item{sds_raw}{\link{data frame} a loaded bedfile with column names annotated.}

\item{df1}{\link{data frame} a dataframe}

\item{df2}{\link{data frame} a dataframe}

\item{along}{\link{1,2} axis along which to zip fasten merge whatever}
}
\value{
a tsv dataframe

a dataframe that is in some intermediate state between bed and tsv and
which needs to be plugged into add_and_fix_columns (another helperfunction) next.

merged dataframe

an error if there is something to complain about. (Päckla Schelln
is schnell aufgmacht :P)
}
\description{
A helperfunction (2/2) for the process of turning a bed to tsv.

A helperfunction (1/2) for the process of turning a bed to tsv.

A helperfunction for the helperfunction. Does some technical dataframe zipping stuff.

QC of the input bedfile.
}
\author{
Wolfram Höps

Wolfram Höps

Unk - someone from the internet.

Wolfram Höps
}
