% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{generate_gt_strings}
\alias{generate_gt_strings}
\title{Generate genotype strings}
\usage{
generate_gt_strings(df, sample_list, haplotype_list)
}
\arguments{
\item{df}{A data frame from the preprocess_data function}

\item{sample_list}{A character vector of unique sample ids}

\item{haplotype_list}{A numeric vector specifying the haplotype options}
}
\value{
A data frame with all the lines for the VCF file
}
\description{
Generate the genotype strings for each unique start and mutation
}
\author{
Wolfram Hoeps
}
