% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{generate_header}
\alias{generate_header}
\title{Generate the header}
\usage{
generate_header(all_lines)
}
\arguments{
\item{all_lines}{A data frame from the generate_gt_strings function}
}
\value{
A character vector with the VCF header
}
\description{
Generate the header for the VCF file
}
\author{
Wolfram Hoeps
}
