% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{get_aln_overlap_in_sector}
\alias{get_aln_overlap_in_sector}
\title{get_aln_overlap_in_sector}
\usage{
get_aln_overlap_in_sector(
  paf,
  x_start,
  y_start,
  x_mid,
  y_mid,
  x_end,
  y_end,
  gp = 10
)
}
\description{
get_aln_overlap_in_sector
}
\author{
Wolfram Hoeps
}
