% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{get_fasta_info}
\alias{get_fasta_info}
\title{Extracts the name and length of each sequence in a fasta file.}
\usage{
get_fasta_info(inputfa, outputinfo)
}
\arguments{
\item{inputfa}{A character string specifying the path to the input fasta file.}

\item{outputinfo}{A character string specifying the path to the output file.}
}
\value{
Nothing is returned by the function, but the sequence names and lengths are saved to the output file.
}
\description{
This function takes an input fasta file and an output file as arguments,
and runs a 'gawk' command to extract the name and length of each sequence
in the fasta file. The extracted information is saved to the output file.
}
\author{
Wolfram Höps
}
