% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{get_subseq}
\alias{get_subseq}
\title{Simple helperfunction to extract part of a fasta into a character.}
\usage{
get_subseq(input_fasta, range = NULL)
}
\arguments{
\item{range}{\link{numeric 2x1 vctor} Sequence range to extract. E.g. range=c(10,100) extracts bases 10 to 100.}

\item{inputfasta}{\link{character/link} link to a single-sequence fasta}
}
\value{
A DNA sequence as a character
}
\description{
Simple helperfunction to extract part of a fasta into a character.
Can be useful e.g. to visualize part of a fasta with an exact dotplot.
}
\author{
Wolfram Höps
}
