% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{gimme_sample_matrix}
\alias{gimme_sample_matrix}
\title{gimme_sample_matrix}
\usage{
gimme_sample_matrix(mode = "diff")
}
\arguments{
\item{mode}{diff/same: should x and y be the same sequence?}
}
\value{
matrix (bitlocus)
}
\description{
Think of this as a function used for development and debugging.
}
\author{
Wolfram Höps
}
