% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{gridlist_to_gridmatrix}
\alias{gridlist_to_gridmatrix}
\title{gridlist_to_gridmatrix}
\usage{
gridlist_to_gridmatrix(grid_list)
}
\arguments{
\item{grid_list}{a dataframe, columns (x,y,z).}
}
\value{
A matrix of the bitplot
}
\description{
Convert gridlist to grimatrix
}
\author{
Wolfram Höps
}
