% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{highlight_region}
\alias{highlight_region}
\title{Highlight a specific region in a given ggplot object}
\usage{
highlight_region(gp, opt)
}
\arguments{
\item{gp}{A ggplot object to highlight a specific region in}

\item{opt}{A list containing necessary options for the highlighting}
}
\value{
A ggplot object with the specified region highlighted
}
\description{
Highlight a specific region in a given ggplot object
}
