% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_logfile.R
\name{init_log_with_def_values}
\alias{init_log_with_def_values}
\title{Initialize log with default values.}
\usage{
init_log_with_def_values()
}
\value{
A named list containing the initialized flags and counters.
}
\description{
This function initializes a named list with the default values for the flags and counters used in the analysis.
}
\author{
Wolfram Hoeps
}
