% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{is_cluttered}
\alias{is_cluttered}
\title{Determine if a bitlocus is good for use in the first place.
If it has many dots around its edges, it is not good and we
need a larger window.}
\usage{
is_cluttered(bitlocus_f, clutter_limit_per_border = 5)
}
\description{
I want to have a second look at this function... is it clean / reasonable?
}
