% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{liftover_coarse}
\alias{liftover_coarse}
\title{liftover_coarse}
\usage{
liftover_coarse(
  seqname,
  start,
  end,
  conversionpaf_link,
  n_probes = 100,
  lenfactor = 1.2,
  whole_chr = F,
  search_mode = "mad",
  refine_runnr = 0
)
}
\description{
Wrapperfunction to find a liftover sequence from one assembly to the other.
Better description TBD.
}
\author{
Wolfram Höps
}
