% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{load_and_prep_paf_for_gridplot_transform}
\alias{load_and_prep_paf_for_gridplot_transform}
\title{load_and_prep_paf_for_gridplot_transform
Called (exclusively) by wrapper_paf_to_bitlocus
Loads and prepares the input paf. Preparations:
Length filter
Start/Endpoint rounding ('compression')
Slope = 1
Transform start / end in negative alignments
Append qmin qmax columns.}
\usage{
load_and_prep_paf_for_gridplot_transform(
  inpaf,
  minlen,
  compression,
  quadrantsize = 1e+05,
  inparam_chunklen = 10000
)
}
\description{
load_and_prep_paf_for_gridplot_transform
Called (exclusively) by wrapper_paf_to_bitlocus
Loads and prepares the input paf. Preparations:
Length filter
Start/Endpoint rounding ('compression')
Slope = 1
Transform start / end in negative alignments
Append qmin qmax columns.
}
\author{
Wolfram Hoeps
}
