% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqplotter.R
\name{make_dotplot}
\alias{make_dotplot}
\title{Main wrapper function for making an exact dotplot. Should not be used for sequences
above 25 kbp.}
\usage{
make_dotplot(
  targetfa,
  queryfa,
  wsize,
  outfile,
  save = T,
  debugmode = F,
  targetrange = NULL,
  queryrange = NULL
)
}
\arguments{
\item{save}{\link{bool} if T, save plot to a file. Otherwise, return.}

\item{seq1link}{\link{character/link} A link to a target fasta file, as character or DNAString}

\item{seq2link}{\link{character/link} A link to a query fasta file, as character or DNAString}
}
\value{
a ggplot2 object or nothing, depending on the 'save' parameter.
}
\description{
Give me two sequences (character of DNAString I think),
and I will return to you a ggplot object with an exact (bp-precise)
dotplot. This relies on the mkDotPlotDataFrame function from 'dotplot' package.
}
\author{
Wolfram Höps
}
