% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_whole_genome_list.R
\name{make_genome_file}
\alias{make_genome_file}
\title{Turn fasta.fai into genome file}
\usage{
make_genome_file(fasta, out_path)
}
\arguments{
\item{out_path}{Path to the output file}

\item{fasta_fai}{Path to the fasta.fai}
}
\value{
Nothing
}
\description{
Turn fasta.fai into genome file
}
