% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_ggplot.R
\name{make_modified_grid_plot}
\alias{make_modified_grid_plot}
\title{Create a modified grid plot}
\usage{
make_modified_grid_plot(res, gridmatrix, outlinks)
}
\arguments{
\item{res}{A data frame with results}

\item{gridmatrix}{A matrix representing the grid}

\item{outlinks}{A list containing output links}
}
\description{
Create a modified grid plot
}
