% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{make_output_folder_structure}
\alias{make_output_folder_structure}
\title{Create empty output folders for NAHRwhals output files}
\usage{
make_output_folder_structure(outdir, sequence_name_output)
}
\arguments{
\item{sequence_name_output}{The output directory path for the files.}
}
\description{
This function creates empty output folders for various intermediate and final files generated by NAHRwhals.
}
