% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_bed_format_and_qc.R
\name{make_sd_partners}
\alias{make_sd_partners}
\title{A helperfunction (1/2) for the process of turning a bed to tsv.}
\usage{
make_sd_partners(sds_raw)
}
\arguments{
\item{sds_raw}{\link{data frame} a loaded bedfile with column names annotated.}
}
\value{
a dataframe that is in some intermediate state between bed and tsv and
which needs to be plugged into add_and_fix_columns (another helperfunction) next.
}
\description{
A helperfunction (1/2) for the process of turning a bed to tsv.
}
\author{
Wolfram Höps
}
