% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_ggplot.R
\name{make_segmented_pairwise_plot}
\alias{make_segmented_pairwise_plot}
\title{make_segmented_pairwise_plot}
\usage{
make_segmented_pairwise_plot(grid_xy, plot_x_y, outlinks)
}
\description{
give me a grid_xy and a plot. I will add segments and save.
This is a bit a quick-and-dirty function; consider to improve.
}
