% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{make_xy_grid_fast}
\alias{make_xy_grid_fast}
\title{Generate a grid of x and y coordinates for a bitlocus plot.}
\usage{
make_xy_grid_fast(paf_df, generations = 1e+05)
}
\arguments{
\item{paf}{a dataframe of alignments which is in paf-like format.}

\item{generations.}{The max number of generations to run this thing. Default infinity.}
}
\value{
A list containing the x and y coordinates of the grid, and a logical indicating whether the grid was successfully generated or not.
}
\description{
This function takes a PAF file, and generates a grid of x and y coordinates for a
bitlocus plot. The grid is generated by first bouncing the start and end points of
each alignment in the PAF file, and then bouncing the resulting points a few more
times. The bounced points are then collected, and the unique points are returned
as the x and y coordinates of the grid.
}
\author{
Wolfram Hoeps
}
