% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{manufacture_output_res_name}
\alias{manufacture_output_res_name}
\title{manufacture_output_res_name}
\usage{
manufacture_output_res_name(seqname_x, start_x, end_x)
}
\description{
Creates an output file name that will be used by other functions.
}
