% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_paf_entries_intraloop.R
\name{merge_paf_entries_intraloop}
\alias{merge_paf_entries_intraloop}
\title{Undocumented helperfunction. Called in the process of stitching together chunked alignments.}
\usage{
merge_paf_entries_intraloop(
  inpaf,
  second_run = F,
  inparam_chunklen = NULL,
  inparam_compression = NULL
)
}
\description{
Undocumented helperfunction. Called in the process of stitching together chunked alignments.
}
\author{
Wolfram Höps
}
