% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_paf.R
\name{merge_rows}
\alias{merge_rows}
\title{Merge paired alignment rows in PAF file}
\usage{
merge_rows(paffile, pairs)
}
\arguments{
\item{paffile}{PAF file data frame with column names}

\item{pairs}{Data frame with columns representing row numbers of paired alignments}
}
\value{
PAF file data frame with merged rows
}
\description{
Merges paired alignment rows in the provided PAF file.
}
\author{
Wolfram Hoeps
}
