% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{modify_gridmatrix}
\alias{modify_gridmatrix}
\title{modify_gridmatrix}
\usage{
modify_gridmatrix(gmx, r1)
}
\arguments{
\item{gmx}{GridMatriX. Row/Colnames don't matter.}

\item{r1}{mutation instruction. Not sure about format right now. TODO.}
}
\description{
Introduce a chain of mutations to gridmatrix.
Is used after mutation simulation to verify/visualize
the result.
}
