% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{mutate_seq}
\alias{mutate_seq}
\title{Introduce mutations to a DNA sequence.}
\usage{
mutate_seq(seq_fasta, sds_tsv, sv_instr_txt, outfasta, outsd, debug = F)
}
\arguments{
\item{seq_fasta}{\link{character/link} A single-sequence fasta file containing the
original sequence, which is to be modified. Should contain SDs.}

\item{sds_tsv}{\link{character/link} A tsv file DESCRIBING the seq_fasta. This information will
be used to update presumed SD position post modification (not working well yet).}

\item{sv_instr_txt}{\link{character/link} A tab-separated file with colums \link{SD-uid, SV}.
SD-uid has to be existing in sds_tsv. SV can be \link{INV, DEL, DUP}. s}

\item{outfasta}{\link{character/link} Outputfile for the modified fasta}

\item{outsd}{\link{character/link} Outputfile for the modified SD tsv file.}
}
\value{
nothing. But output files written.
}
\description{
This is sitting at the core of the nahrchainer module.
}
\author{
Wolfram Höps
}
