% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nahrwhals.R
\name{nahrwhals}
\alias{nahrwhals}
\title{NW main}
\usage{
nahrwhals(
  ref_fa,
  asm_fa,
  regions = NULL,
  genome_y_fa_mmi = "default",
  anntrack = FALSE,
  logfile = "res/res.tsv",
  res_bedfile = "res/res_all.bed",
  samplename_x = "Fasta_x",
  samplename_y = "Fasta_y",
  compare_full_fastas = FALSE,
  plot_only = FALSE,
  self_plots = TRUE,
  plot_xy_segmented = TRUE,
  eval_th = 98,
  depth = 3,
  chunklen = "default",
  minlen = "default",
  compression = "default",
  max_size_col_plus_rows = 250,
  max_n_alns = 150,
  use_paf_library = FALSE,
  conversionpaf_link = FALSE,
  xpad = 1,
  plot_minlen = 350,
  maxlen_refine = 1e+10,
  n_tests = 10,
  n_max_testchunks = 5,
  baseline_log_minsize_min = 8,
  baseline_log_minsize_max = FALSE,
  discovery_exact = FALSE,
  hltrack = FALSE,
  hllink = FALSE,
  aln_pad_factor = 1,
  debug = FALSE,
  clean_after_yourself = FALSE,
  testrun_std = FALSE,
  testrun_fullfa = FALSE,
  threads = 1,
  noclutterplots = T,
  maxdup = 2,
  minreport = 0.98,
  init_width = 1000,
  minimap_cores = 1
)
}
\description{
NW main
}
