% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{filter_paf_to_main_region}
\alias{filter_paf_to_main_region}
\title{filter_paf_to_main_region
Script to filter down a whole-genome paf to the 'best matching' region.
query should be short sequence fragments (1kb, 10kb, ...) of the roi.}
\usage{
filter_paf_to_main_region(paflink, outpaflink)
}
\description{
filter_paf_to_main_region
Script to filter down a whole-genome paf to the 'best matching' region.
query should be short sequence fragments (1kb, 10kb, ...) of the roi.
}
\author{
Wolfram Hoeps
}
