% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_to_bed.R
\name{paf_to_bed}
\alias{paf_to_bed}
\title{Convert a trimmed-down paf to bed? Not sure either about this.}
\usage{
paf_to_bed(inpaf)
}
\arguments{
\item{inpaf}{\link{data frame} a link to the inpaf to be converted to bed.}
}
\value{
a bed-file style dataframe.
}
\description{
Convert a trimmed-down paf to bed? Not sure either about this.
}
\author{
Wolfram Höps
}
