% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_to_bed.R
\name{paf_to_sd_paf}
\alias{paf_to_sd_paf}
\title{Helperfunction (1/2) for paf_write_bed.}
\usage{
paf_to_sd_paf(inpaf_link)
}
\arguments{
\item{inpaf_link}{A link to the inpaf to be converted to bed.}
}
\value{
a filtered pseudo-paf file. Gotta double-check this.
}
\description{
Cuts down a paf to entries where query > target. Not 100\% sure
what is going on here.
}
\author{
Wolfram Höps
}
