% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{pafdotplot_make}
\alias{pafdotplot_make}
\title{Generate a Dotplot from a PAF File with Optional SD Highlight Annotation}
\usage{
pafdotplot_make(
  inpaf_link,
  outplot_link,
  min_align = 11,
  min_query_aln = 11,
  keep_ref = 10000,
  similarity = TRUE,
  h_lines = TRUE,
  interactive = FALSE,
  plot_size = 10,
  on_target = TRUE,
  v = FALSE,
  hllink = FALSE,
  hltype = FALSE,
  hlstart = NULL,
  hlend = NULL,
  save = TRUE,
  minsdlen = 5000,
  anntrack = FALSE,
  x_seqname = NULL,
  x_start = NULL,
  x_end = NULL,
  hltrack = NULL,
  aln_type_xx_yy_xy = "xy",
  samplename_x = "seq_x",
  samplename_y = "seq_y"
)
}
\arguments{
\item{inpaf_link}{Character/link: Path to the PAF file to be plotted.}

\item{outplot_link}{Character/link: Path to the output dotplot PDF to be saved.}

\item{min_align}{Numeric: Minimum alignment length (default: 11).}

\item{min_query_aln}{Numeric: Minimum query alignment length (default: 11).}

\item{keep_ref}{Numeric: Number of sequences to keep for plotting.}

\item{similarity}{Logical: Indicate sequence similarity by color code (default: TRUE).}

\item{h_lines}{Logical: Plot horizontal lines (default: TRUE).}

\item{interactive}{Logical: Use interactive dotplot (Dotplotly)? Untested, keep at FALSE.}

\item{plot_size}{Numeric: Plot size in inches.}

\item{on_target}{Logical: Unknown parameter, default: TRUE.}

\item{v}{Logical: Unknown parameter, default: FALSE.}

\item{hllink}{Character/link: Path to an SD annotation file (BED, TSV, or PAF) to include as highlights in the plot.}

\item{hltype}{Character: Filetype of hllink. Can be 'NULL', 'bed', 'tsv', 'paf'.}

\item{hlstart}{Numeric: Optional start position for highlights.}

\item{hlend}{Numeric: Optional end position for highlights.}

\item{save}{Logical: Save the dotplot to the specified output path (default: TRUE).}

\item{minsdlen}{Numeric: Minimum length for SD annotations (default: 5000).}

\item{anntrack}{Logical: Include annotation tracks (default: FALSE).}

\item{x_seqname}{Character: Sequence name for X-axis.}

\item{x_start}{Numeric: Start position for X-axis.}

\item{x_end}{Numeric: End position for X-axis.}

\item{hltrack}{Logical: Include highlight track (default: NULL).}

\item{aln_type_xx_yy_xy}{Character: Alignment type (default: 'xy').}

\item{samplename_x}{The name of the sequence to appear on the x axis. (default: 'seq_x')}

\item{samplename_y}{The name of the sequence to appear on the y axis. (default: 'seq_y')}
}
\value{
The generated dotplot. If 'save' is TRUE, the dotplot is saved as a PDF.
}
\description{
This function generates a dotplot from a PAF file and saves it as a PDF.
It can also include an optional SD highlight annotation file for enhanced visualization.
}
\author{
Wolfram Höps
}
