% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{plot_alignments}
\alias{plot_alignments}
\title{Plot alignments}
\usage{
plot_alignments(alignments, opt)
}
\arguments{
\item{alignments}{A data frame containing alignment information}

\item{opt}{A list of options for the plot}
}
\description{
Plot alignments
}
