% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{plot_helper_debug}
\alias{plot_helper_debug}
\title{Debugging function for plotting points and PAF data}
\usage{
plot_helper_debug(paf, points_overall)
}
\arguments{
\item{paf}{A data frame of alignments in .paf-like format.}

\item{points_overall}{A data frame of points.}
}
\description{
This function takes a data frame of alignments in .paf-like format, and a data frame
of points, and generates a plot to help with debugging.
}
