% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_search_tools.R
\name{plot_matrix}
\alias{plot_matrix}
\title{A function for plotting a matrix}
\usage{
plot_matrix(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\value{
none, but image is plotted
}
\description{
Found on stackexchange. Low-yield matrix plot. For quick/dirty
visualizations.
}
\author{
unknown (http://www.phaget4.org/R/image_matrix.html)
}
