% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_ggplot.R
\name{plot_matrix_ggplot}
\alias{plot_matrix_ggplot}
\title{Plot_matrix_ggplot}
\usage{
plot_matrix_ggplot(data_frame_xyz)
}
\arguments{
\item{data_frame_xyz}{a dataframe with three coordinate columns (x,y,z)}
}
\description{
Make a plot of a dataframe matrix.
}
\author{
Wolfram Hoeps
}
