% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{plot_paf}
\alias{plot_paf}
\title{plot_paf
This is another, MUCH simplified version of plotting a paf.
Optionally, gridlines can be plotted, too.}
\usage{
plot_paf(
  paf_df,
  gridlines.x = NULL,
  gridlines.y = NULL,
  wiggle = 0,
  plot_gridlines = T,
  linewidth = 1
)
}
\description{
plot_paf
This is another, MUCH simplified version of plotting a paf.
Optionally, gridlines can be plotted, too.
}
\author{
Wolfram Hoeps
}
