% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_ggplot.R
\name{pmsqrt}
\alias{pmsqrt}
\title{Compute the signed square root}
\usage{
pmsqrt(x)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
The signed square root of the input value(s)
}
\description{
Compute the signed square root
}
