% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess input data}
\usage{
preprocess_data(input_file, res_max_threshold = 0.98)
}
\arguments{
\item{input_file}{A character string specifying the path to the input file}

\item{res_max_threshold}{A numeric value specifying the res_max threshold for assigning 'UNK' to mut_max}
}
\value{
A preprocessed data frame
}
\description{
Remove alt samples, replace mut_max by 'UNK' where res_max < threshold, extract sample_id and phase from sample
}
\author{
Wolfram Hoeps
}
