% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{query_config}
\alias{query_config}
\title{A wrapper for extracting values from the config.}
\usage{
query_config(param)
}
\description{
A wrapper for extracting values from the config.
}
\author{
Wolfram Höps
}
