% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{randDNASeq}
\alias{randDNASeq}
\title{Create a simple random sequence.}
\usage{
randDNASeq(n, gcfreq, seed = 1234)
}
\arguments{
\item{n}{\link{numeric} length of desired sequence (bp)}

\item{gcfreq}{\link{character/link} desired GC frequency.}
}
\value{
A character vector of a random DNA sequence.
}
\description{
Simple function to create a random string of ACGT.
}
\author{
Wolfram Höps
}
