% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{rc}
\alias{rc}
\title{https://rdrr.io/cran/insect/src/R/complement.R
Reverse complement DNA in character string format.}
\usage{
rc(z)
}
\arguments{
\item{z}{a vector of DNA sequences in upper case character string format.}
}
\value{
a vector of DNA sequences as upper case character strings.
}
\description{
This function reverse complements a DNA sequence or vector of DNA
sequences that are stored as character strings.
}
\details{
This function accepts only DNA sequences in concatenated character
string format, see \code{\link[ape]{complement}} in the \code{\link[ape]{ape}}
package for "DNAbin" input objects, and \code{\link[seqinr]{comp}} in the
\code{\link[seqinr]{seqinr}} package for when the input object is a character
vector.
}
\examples{
rc("TATTG")
}
\author{
Shaun Wilkinson
}
