% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{read_and_prep_paf}
\alias{read_and_prep_paf}
\title{read_and_prep_paf}
\usage{
read_and_prep_paf(paflink)
}
\arguments{
\item{paflink}{The path to the PAF file to read.}
}
\description{
Reads in a PAF file and renames the columns to match the PAF standard column names.
}
\author{
Wolfram Hoeps
}
