% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates_triple.R
\name{remove_duplicates_triple}
\alias{remove_duplicates_triple}
\title{remove_duplicates_triple}
\usage{
remove_duplicates_triple(grid_list_f)
}
\description{
Remove duplicates TRIPLE. First, remove obvious duplicates (all 3 same)
Then, identify those with same x/y but different z. The get 0.
Finally, the second step has produced new duplicates, so be mean again.
Used to 'purify' / 'simplify' a grid_list
}
\author{
Wolfram Hoeps
}
