% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltools.R
\name{repeat_col_in_matrix}
\alias{repeat_col_in_matrix}
\title{repeat_col_in_matrix}
\usage{
repeat_col_in_matrix(x_val, n_val)
}
\arguments{
\item{x_val}{The column to be repeated.}

\item{n_val}{The number of times the column should be repeated.}
}
\value{
A matrix with the repeated columns.
}
\description{
This function takes a column and repeats it n times in a matrix.
}
\author{
Wolfram Hoeps
}
