% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaltools.R
\name{repeat_row_in_matrix}
\alias{repeat_row_in_matrix}
\title{repeat_row_in_matrix}
\usage{
repeat_row_in_matrix(x_val, n_val)
}
\arguments{
\item{x_val}{The row to be repeated.}

\item{n_val}{The number of times the row should be repeated.}
}
\value{
A matrix with the repeated rows.
}
\description{
This function takes a row and repeats it n times in a matrix.
}
\author{
Wolfram Hoeps
}
