% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{return_diag_values_new}
\alias{return_diag_values_new}
\title{Diagonalize.}
\usage{
return_diag_values_new(matrix, fraction = 0.05)
}
\arguments{
\item{matrix}{a matrix typically representing a bitlocus}

\item{fraction, }{how 'thick' is the piece that we are reporting (0 = nothing, 1 = whole matrix)}
}
\description{
Return values near the diagonal of a matrix.
}
