% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{run_minimap2}
\alias{run_minimap2}
\title{Submit a system command to run minimap2}
\usage{
run_minimap2(targetfasta, queryfasta, outpaf, params, nthreads = 4)
}
\arguments{
\item{targetfasta}{\link{character/link} link to the 'target' single-sequence fasta (sometimes reference, e.g. chm13.)}

\item{queryfasta}{\link{character/link} link to the 'query' fasta. Can be single or multi-fasta}

\item{outpaf}{\link{character/link} Path to the output paffile to be written.}

\item{minimap2loc}{\link{character/link} link to minimap2 binary.}
}
\value{
nothing. Only output files written.
}
\description{
This is a helperfunction to run minimap2. Also check out:
https://github.com/PacificBiosciences/pbmm2/ . Minimap2 parameters:
-k   k-mer size (no larger than 28). \link{-1}
-w   Minimizer window size. \link{-1}
-u   Disable homopolymer-compressed k-mer (compression is active for SUBREAD & UNROLLED presets).
-A   Matching score. \link{-1}
-B   Mismatch penalty. \link{-1}
-z   Z-drop score. \link{-1}
-Z   Z-drop inversion score. \link{-1}
-r   Bandwidth used in chaining and DP-based alignment. \link{-1}
-g   Stop chain enlongation if there are no minimizers in N bp. \link{-1}
#'
}
\author{
Wolfram Höps
}
