% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_find_parameters.R
\name{run_n_tests}
\alias{run_n_tests}
\title{TODO: describe}
\usage{
run_n_tests(
  inpaf,
  min_,
  max_,
  res,
  n_tests,
  n_max_alns,
  max_size_col_plus_rows
)
}
\description{
TODO: describe
}
