% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_bed_format_and_qc.R
\name{run_qc_bed_df}
\alias{run_qc_bed_df}
\title{QC of the input bedfile.}
\usage{
run_qc_bed_df(sds_raw)
}
\arguments{
\item{sds_raw}{\link{data frame} a loaded bedfile with column names annotated.}
}
\value{
an error if there is something to complain about. (Päckla Schelln
is schnell aufgmacht :P)
}
\description{
QC of the input bedfile.
}
\author{
Wolfram Höps
}
