% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{run_steepest_descent_one_layer_down}
\alias{run_steepest_descent_one_layer_down}
\title{run_steepest_descent_one_layer_down}
\usage{
run_steepest_descent_one_layer_down(bitlocus, res_df, starting_depth = 1)
}
\description{
Helperfunction to run a tree search in steepest_descent mode.
}
