% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_to_bed.R
\name{sd_to_bed}
\alias{sd_to_bed}
\title{Convert a (normal, double-entry-) tsv file to bed}
\usage{
sd_to_bed(sdlink, outbedfile = NULL)
}
\arguments{
\item{sdlink}{\link{character/link} a link to the input sd file to be converted to bed.}

\item{outbedfile}{\link{character/link/NULL} a link to the output bedfile. If null, a
bed-like dataframe is returned}
}
\value{
a bed-file style dataframe if outbedfile == NULL
}
\description{
Convert a (normal, double-entry-) tsv file to bed
}
\author{
Wolfram Höps
}
