% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmodder.R
\name{sds_document_deletion}
\alias{sds_document_deletion}
\title{Document the introduction of a DEL in the SD tsv dataframe.}
\usage{
sds_document_deletion(sds, sv)
}
\arguments{
\item{sds}{\link{data.frame} A dataframe with tsv SD information}

\item{sv}{\link{data.frame} A data framewith columns \link{SD-uid, SV}.
SD-uid has to be existing in sds_tsv. SV can be \link{INV, DEL, DUP}}
}
\value{
a mutated sds dataframe.
}
\description{
This function is typically called by the carry_out_del function.
ALSO the function is highly experimental and probably not working.
We recommend not rely on it at this point.
}
\author{
Wolfram Höps
}
