% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder.R, R/seqbuilder_functions.R
\name{model_sds}
\alias{model_sds}
\alias{add_snps}
\alias{simulate_seq}
\alias{randDNASeq}
\alias{wrapper_dotplot_with_alignment_fast}
\title{Introduce repeats to a simulated sequence}
\usage{
model_sds(seq_base, sds, uid)

add_snps(seq, similarity, seed = 1234)

simulate_seq(seqlen, sdfile, outfasta, debugmode = F)

randDNASeq(n, gcfreq, seed = 1234)

wrapper_dotplot_with_alignment_fast(
  seqname,
  start,
  end,
  genome_x_fa,
  genome_y_fa,
  subseqfasta_x,
  subseqfasta_y,
  conversionpaf_link,
  outpaf_link,
  chunklen = 10000,
  factor = 0.5
)
}
\arguments{
\item{seq_base}{A DNA sequence, as character}

\item{sds}{A table including columns 'chromStart', 'chromEnd', 'uid', 'strand'}

\item{uid}{Name of the SD pair to be created. Must appear (exactly) twice in sds.}

\item{seq}{DNA sequence, as 'character'}

\item{similarity}{How similar should the result be to the input (fraction of bases, \link{0-1})}

\item{seqlen}{\link{numeric} length of simulated sequence in bp}

\item{sdfile}{\link{character} A link to an SD file of the form of SD_hg38.tsv tables.}

\item{outfasta}{\link{character} a link to the outputfasta that will be saved}

\item{n}{\link{numeric} length of desired sequence (bp)}

\item{gcfreq}{\link{character/link} desired GC frequency.}
}
\value{
A character vector of a random DNA sequence.

A character vector of a random DNA sequence.
}
\description{
This function takes a DNA sequence and an SD instruction (.tsv format),
and introduces this SD.
Also calls add_snps function, to introduce snps in the SD pair,
to create the similarity specified in the input tsv.

Mutate a proportion of bases in a DNA sequence. Can be used to
simulate evolutionary divergence of two SD pairs.

Simple function to create a random string of ACGT.

Simple function to create a random string of ACGT.
}
\examples{

Some example function call or so.

}
\author{
Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps

Wolfram Höps
}
