% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{shorten_fasta}
\alias{shorten_fasta}
\title{helperfunction to shorten a fasta file.}
\usage{
shorten_fasta(infasta, outfasta, range)
}
\description{
helperfunction to shorten a fasta file.
}
\author{
Nicholas Hathaway
}
