% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{shred_seq}
\alias{shred_seq}
\title{Chunkify query fasta}
\usage{
shred_seq(infasta, outfasta_chunk, chunklen)
}
\arguments{
\item{infasta}{\link{character/link} single-seq fasta to be chopped}

\item{outfasta_chunk}{\link{character/link} output chopped multi-seq fasta.}

\item{chunklen}{\link{numeric} length of sequence chunks in bp}

\item{scriptloc}{\link{character/link} link to shred.ss from bbmap.}
}
\value{
nothing. Only output files written.
}
\description{
This is a helperfunction calling an external script to
chop a query sequence into chunks.
}
\author{
Wolfram Höps
}
