% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{shred_seq_bedtools}
\alias{shred_seq_bedtools}
\title{Chunkify query fasta}
\usage{
shred_seq_bedtools(infasta, outfasta_chunk, chunklen, params)
}
\arguments{
\item{infasta}{\link{character/link} single-seq fasta to be chopped}

\item{outfasta_chunk}{\link{character/link} output chopped multi-seq fasta.}

\item{chunklen}{\link{numeric} length of sequence chunks in bp}
}
\value{
nothing. But output files written.
}
\description{
This is a helperfunction calling bedtools to
chop a query sequence into chunks.
}
\author{
Wolfram Höps
}
