% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{shred_seq_bedtools_multifasta}
\alias{shred_seq_bedtools_multifasta}
\title{shred_seq_bedtools_multifasta}
\usage{
shred_seq_bedtools_multifasta(infasta, outfasta_chunk, chunklen, params)
}
\arguments{
\item{infasta}{A link to a multi-seq fasta to be chopped}

\item{outfasta_chunk}{A link to the output path for a chopped multi-seq fasta.}

\item{chunklen}{length of sequence chunks in bp}

\item{params}{a list with all NAHRwhals parameters}
}
\value{
nothing. But output files written.
}
\description{
This is a helperfunction calling bedtools to
chop a query sequence into chunks.
}
\author{
Wolfram Hoeps
}
