% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder.R
\name{simulate_seq}
\alias{simulate_seq}
\title{Create a simulated DNA sequence, containing SDs of desired similarity}
\usage{
simulate_seq(seqlen, sdfile, outfasta, debugmode = F)
}
\arguments{
\item{seqlen}{\link{numeric} length of simulated sequence in bp}

\item{sdfile}{\link{character} A link to an SD file of the form of SD_hg38.tsv tables.}

\item{outfasta}{\link{character} a link to the outputfasta that will be saved}
}
\description{
Create a simulated DNA sequence, containing SDs of desired similarity
}
\author{
Wolfram Höps
}
