% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{transform_res_new_to_old}
\alias{transform_res_new_to_old}
\title{transform_res_new_to_old}
\usage{
transform_res_new_to_old(res_df_f)
}
\description{
This is a helperfunction transforming between different ways of representing the results of a tree search.
Further documentation required - for now, just don't touch it :).
}
