% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2t_names_translate.R
\name{translate_t2t_chr_to_readable}
\alias{translate_t2t_chr_to_readable}
\title{Translate T2T Chromosome Names to Readable Format}
\usage{
translate_t2t_chr_to_readable(chrname)
}
\arguments{
\item{chrname}{A character string representing the cryptic chromosome name
to be translated.}
}
\value{
A character string with the translated chromosome name in a more
readable format. If the input name is not found in the T2T dataset, it
returns the original input name.
}
\description{
This function translates cryptic chromosome names in the
T2T dataset to more readable and user-friendly names. If a given chromosome
name is not found in the T2T dataset, the function returns the original input name.
}
