% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_compare_input_output.R
\name{turn_intervals_to_vector_0_1}
\alias{turn_intervals_to_vector_0_1}
\title{A helperfunction (1/1) for compare_sd_paf_input_output.}
\usage{
turn_intervals_to_vector_0_1(intervals, maxlen)
}
\arguments{
\item{intervals}{\link{data.frame} A data frame with interval columns \link{start, stop}}

\item{maxlen}{\link{character/link/NULL} Total length of the vector to create. Should
be >= max(invervals$stop)}
}
\value{
a vector of length maxlen, with 0/1 depending on coverage of intervals.
}
\description{
takes a list of intervals, and a max sequence length. Returns
a 0-1 vector where 1 is spaces covered by an SD. This is a quite inefficient
algorithm, but the part is not extremely time-sensitive.
}
\author{
Wolfram Höps
}
