% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juliasolver.R
\name{turn_mut_max_into_svdf}
\alias{turn_mut_max_into_svdf}
\title{turn_mut_max_into_svdf
A function to correct the coordinates of the mutations in the mut_max column}
\usage{
turn_mut_max_into_svdf(t, correction = T, return_string = F)
}
\description{
turn_mut_max_into_svdf
A function to correct the coordinates of the mutations in the mut_max column
}
