% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{wrapper_aln_and_analyse}
\alias{wrapper_aln_and_analyse}
\title{wrapper_aln_and_analyse}
\usage{
wrapper_aln_and_analyse(params)
}
\arguments{
\item{params}{a single list containing all input information.}
}
\description{
The main wrapper function for running NAHRwhals.
This is where all the information flow together.
}
