% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{wrapper_dotplot_with_alignment_fast}
\alias{wrapper_dotplot_with_alignment_fast}
\title{General purpose wrapper}
\usage{
wrapper_dotplot_with_alignment_fast(
  seqname,
  start,
  end,
  genome_x_fa,
  genome_y_fa,
  subseqfasta_x,
  subseqfasta_y,
  conversionpaf_link,
  outpaf_link,
  chunklen = 10000,
  factor = 0.5
)
}
\arguments{
\item{n}{\link{numeric} length of desired sequence (bp)}

\item{gcfreq}{\link{character/link} desired GC frequency.}
}
\value{
A character vector of a random DNA sequence.
}
\description{
Simple function to create a random string of ACGT.
}
\author{
Wolfram Höps
}
