% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{write_results}
\alias{write_results}
\title{write_results}
\usage{
write_results(res, outlinks, params)
}
\description{
Write results of a tree search to an output file. Uses the global 'log' variable.
}
