% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Write VCF file}
\usage{
write_vcf(
  input_file,
  output_file,
  res_max_threshold = 0.98,
  sort = F,
  ps_colname = "start"
)
}
\arguments{
\item{input_file}{A character string specifying the path to the input file}

\item{output_file}{A character string specifying the path to the output VCF file}

\item{res_max_threshold}{A numeric value specifying the res_max threshold for assigning 'UNK' to mut_max}
}
\description{
Create VCF file from input data and save it
}
\author{
Wolfram Hoeps
}
