% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_find_extract_2.R
\name{write_x_y_sequences}
\alias{write_x_y_sequences}
\title{write_x_y_sequences}
\usage{
write_x_y_sequences(
  seqname_x,
  start_x_pad,
  end_x_pad,
  genome_x_fa_subseq,
  genome_y_fa_subseq,
  genome_y_fa,
  conversionpaf_link,
  outlinks,
  params,
  extract_only = F
)
}
\description{
This was previously part of the wrapper_aln_and_analyse.
However, over time, this grew into more and more complex code,
So it is for sure worth its own function by now.
}
