% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{writeFasta}
\alias{writeFasta}
\title{Helperfunction to save a fasta file.}
\usage{
writeFasta(data, filename)
}
\arguments{
\item{data}{A data frame with column 'name' and 'seq'}

\item{filename}{An output filename.}
}
\value{
nothing. Only output files written.
}
\description{
a simple function that takes a data.frame that has a column name
and seq and writes a fasta file from it. Taken from
https://bootstrappers.umassmed.edu/guides/main/r_writeFasta.html.
}
\author{
Nicholas Hathaway
}
