% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_bed_format_and_qc.R
\name{zipFastener}
\alias{zipFastener}
\title{A helperfunction for the helperfunction. Does some technical dataframe zipping stuff.}
\usage{
zipFastener(df1, df2, along = 2)
}
\arguments{
\item{df1}{\link{data frame} a dataframe}

\item{df2}{\link{data frame} a dataframe}

\item{along}{\link{1,2} axis along which to zip fasten merge whatever}
}
\value{
merged dataframe
}
\description{
A helperfunction for the helperfunction. Does some technical dataframe zipping stuff.
}
\author{
Unk - someone from the internet.
}
