% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{add_slope_intercept_info}
\alias{add_slope_intercept_info}
\title{add_slope_intercept_info}
\usage{
add_slope_intercept_info(
  vector_df,
  xstart = "tstart",
  xend = "tend",
  ystart = "qstart",
  yend = "qend"
)
}
\description{
add_slope_intercept_info
}
\author{
Wolfram Hoeps
}
