% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{awk_edit_paf}
\alias{awk_edit_paf}
\title{Submit a system command to run awk to change sequence chunk names}
\usage{
awk_edit_paf(inpaf, outpaf, params)
}
\arguments{
\item{inpaf}{A character/link to input PAF file name or link.}

\item{outpaf}{A character/link to output PAF file name or link.}

\item{params}{A named list containing the parameters for running the awk command. The list must contain an element
named \code{awkscript_paf} that specifies the path to the awk script file to be used.}
}
\value{
Nothing. The function only writes the output file.
}
\description{
This is a helper function that submits a system command to run awk to change sequence chunk names in a PAF file.
}
\author{
Wolfram Hoeps
}
