% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bresenham.R
\name{bresenham}
\alias{bresenham}
\title{Bresenham's Integer Line Drawing Algorithm}
\usage{
bresenham(
  x_in,
  y_in,
  gridpoints_x,
  gridpoints_y,
  griddiffs_x,
  griddiffs_y,
  x_grididx,
  y_grididx
)
}
\arguments{
\item{x_in}{numeric vector of the x coordinates of the points to be joined.}

\item{y_in}{numeric vector of the y coordinates of the points to be joined.}

\item{gridpoints_x}{numeric vector of the x coordinates of the grid points.}

\item{gridpoints_y}{numeric vector of the y coordinates of the grid points.}

\item{griddiffs_x}{numeric vector of the differences between the x coordinates}

\item{griddiffs_y}{numeric vector of the differences between the y coordinates}
}
\value{
A list of length 2 with \code{integer} \code{x,y} coordinates connecting
the vertices
}
\description{
Generate integer x,y points between vertices by Bresenham's algorithm.
}
