% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bresenham.R
\name{bresenham}
\alias{bresenham}
\title{Bresenham's Integer Line Drawing Algorithm}
\usage{
bresenham(x, y = NULL, gridpoints_x, gridpoints_y, close = TRUE, debug = F)
}
\arguments{
\item{x}{numeric vector of the x coordinates of the points to be joined.}

\item{y}{numeric vector of the y coordinates of the points to be joined.}

\item{gridpoints_x}{numeric vector of the x coordinates of the grid points.}

\item{gridpoints_y}{numeric vector of the y coordinates of the grid points.}

\item{close}{logical value indicating if the points form a closed polygon.}

\item{debug}{logical value indicating whether to enter debugging mode or not.}
}
\value{
A list of length 2 with \code{integer} \code{x,y} coordinates connecting
the vertices
}
\description{
Generate integer x,y points between vertices by Bresenham's algorithm.
}
