% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{define_output_files}
\alias{define_output_files}
\title{Define a hell lot of output files.
There must be a more elegant way to do this btw}
\usage{
define_output_files(sequence_name_output, samplename)
}
\description{
Define a hell lot of output files.
There must be a more elegant way to do this btw
}
