% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{determine_compression}
\alias{determine_compression}
\title{determine_compression}
\usage{
determine_compression(start, end)
}
\description{
Compression parameter for segmenting the dotplot.
We choose this based on the length of the sequence.
}
