% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_machinery.R
\name{dfs}
\alias{dfs}
\title{DFS caller function}
\usage{
dfs(bitlocus, maxdepth = 3, increase_only = F, earlystop = Inf, history = NULL)
}
\arguments{
\item{bitlocus}{A matrix representing the search space.}

\item{maxdepth}{The maximum depth of the search tree.}

\item{increase_only}{A boolean indicating whether only increasing mutations are allowed.}

\item{earlystop}{The maximum number of evaluations before the search is terminated.}

\item{history}{An optional argument representing previous search history.}
}
\value{
A list containing the search history and the evaluation results.
}
\description{
This function serves as the gateway to entering the DFS process. It calls the dfsutil function recursively
to explore possible paths within the search space.
}
\author{
Wolfram Höps
}
