% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_machinery.R
\name{dfsutil}
\alias{dfsutil}
\title{DFS workhorse function
TODO: description}
\usage{
dfsutil(
  visited,
  pair,
  pairhash,
  mutator,
  depth,
  maxdepth = 3,
  pairhistory = NULL,
  df_output = NULL,
  increase_only = NULL,
  orig_symm = 1,
  last_eval = 0,
  earlystop = Inf
)
}
\description{
DFS workhorse function
TODO: description
}
