% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{enlarge_interval_by_factor}
\alias{enlarge_interval_by_factor}
\title{enlarge_interval_by_factor}
\usage{
enlarge_interval_by_factor(
  start,
  end,
  factor,
  seqname_f = NULL,
  conversionpaf_f = NULL,
  log_collection = NULL
)
}
\description{
input start and end coordinates and a factor by which to enlarge. E.g. 40-50 (len 10) and factor 2 will result in 35-55 (len 20).
Makes sure no negative values are created.
}
\author{
Wolfram Höps
}
