% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{extract_subseq_bedtools}
\alias{extract_subseq_bedtools}
\title{extract_subseq_bedtools}
\usage{
extract_subseq_bedtools(infasta, seqname, start, end, outfasta)
}
\description{
extract_subseq_bedtools
}
\author{
Wolfram Hoeps
}
