% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{find_coords_extrapolated}
\alias{find_coords_extrapolated}
\title{find_coords_extrapolated}
\usage{
find_coords_extrapolated(
  liftover_coords,
  cpaf,
  winner_chr,
  start,
  end,
  liftover_run
)
}
\description{
find_coords_extrapolated
}
\author{
Wolfram Hoeps
}
