% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extraction_functions.R
\name{find_punctual_liftover}
\alias{find_punctual_liftover}
\title{find_punctual_liftover
Translates a single point from one assembly to another.
Uses the pre-computed conversion-paf.}
\usage{
find_punctual_liftover(cpaf, pointcoordinate, chrom)
}
\description{
find_punctual_liftover
Translates a single point from one assembly to another.
Uses the pre-computed conversion-paf.
}
