% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{find_threshold}
\alias{find_threshold}
\title{Which percentage do we call 'solved'?
To account for alignment incongruencies, we say that a puzzle is solved
If it deviates at most by 1 'unit'.}
\usage{
find_threshold(bitlocus_f, nmut)
}
\description{
I want to have a second look at this funciton... is it clean / reasonable?
}
