% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlist_to_gridmatrix.R
\name{gridlist_to_gridmatrix}
\alias{gridlist_to_gridmatrix}
\title{gridlist_to_gridmatrix}
\usage{
gridlist_to_gridmatrix(grid)
}
\arguments{
\item{grid}{a dataframe, columns (x,y,z).}
}
\value{
A matrix of the bitplot
}
\description{
Convert gridlist to grimatrix
}
\author{
Wolfram Hoeps
}
