% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{is_cluttered}
\alias{is_cluttered}
\title{is_cluttered}
\usage{
is_cluttered(bitlocus_f, clutter_limit_per_border = 5)
}
\arguments{
\item{bitlocus_f}{The matrix representing the bitlocus.}

\item{clutter_limit_per_border}{The maximum number of dots allowed per border.}
}
\value{
TRUE if the bitlocus is cluttered, FALSE otherwise.
}
\description{
This function determines if a bitlocus is too cluttered to be analysed/considered further
}
