% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqbuilder_functions.R
\name{make_chunked_minimap_alnment}
\alias{make_chunked_minimap_alnment}
\title{Perform Chunked Minimap2 Alignment and Generate Dotplot}
\usage{
make_chunked_minimap_alnment(
  params,
  targetfasta,
  queryfasta,
  outpaf,
  chunklen = 1000,
  keep_ref = 10000,
  minsdlen = 2000,
  plot_size = 10,
  saveplot = T,
  savepaf = T,
  quadrantsize = 1e+05,
  hllink = F,
  hltype = F,
  hlstart = NULL,
  hlend = NULL,
  targetrange = NULL,
  queryrange = NULL,
  anntrack = F,
  x_seqname = NULL,
  x_start = NULL,
  x_end = NULL,
  hltrack = NULL,
  onlypafreturn = F,
  aln_type_xx_yy_xy = "xy"
)
}
\arguments{
\item{params}{a list of NAHRwhals paramters}

\item{targetfasta}{A character string containing the path to the 'target' single-sequence FASTA file (e.g., chm13).}

\item{queryfasta}{A character string containing the path to the 'query' single-sequence FASTA file.}

\item{outpaf}{A character string containing the path to the output PAF file to be written.}

\item{chunklen}{A numeric value indicating the length of sequence chunks to split the query into (default = 1000 bp).}

\item{keep_ref}{A numeric value indicating the number of alignments to keep for plotting (default = 10000).}

\item{minsdlen}{A numeric value; purpose not entirely clear.}

\item{plot_size}{A numeric value indicating the plot size in inches (default = 10).}

\item{saveplot}{A logical value indicating whether or not to save the plot (default = TRUE).}

\item{savepaf}{A logical value indicating whether or not to save the PAF file (default = TRUE).}

\item{quadrantsize}{A numeric value indicating the quadrant size (default = 100000).}

\item{hllink}{A character string containing the path to an SD annotation file (BED, TSV, or PAF) to include as highlights in the plot (default = FALSE).}

\item{hltype}{A character string indicating the filetype of hllink. Can be 'NULL', 'bed', 'tsv', or 'paf' (default = FALSE).}

\item{hlstart}{An integer value indicating the start position of the highlight region (default = NULL).}

\item{hlend}{An integer value indicating the end position of the highlight region (default = NULL).}

\item{targetrange}{A numeric vector containing the start and end positions of the target sequence to use (default = NULL).}

\item{queryrange}{A numeric vector containing the start and end positions of the query sequence to use (default = NULL).}

\item{anntrack}{A logical value indicating whether or not to include an annotation track (default = FALSE).}

\item{x_seqname}{A character string indicating the name of the reference sequence for the annotation track (default = NULL).}

\item{x_start}{A numeric value indicating the start position of the reference sequence for the annotation track (default = NULL).}

\item{x_end}{A numeric value indicating the end position of the reference sequence for the annotation track (default = NULL).}

\item{hltrack}{A numeric value indicating the track number to use for highlighting (default = NULL).}

\item{onlypafreturn}{A logical value indicating whether or not to only return the PAF file (default = FALSE).}

\item{aln_type_xx_yy_xy}{A character string indicating the orientation of the alignment (default = 'xy').}
}
\value{
This function has no return value. Output files are written to the specified paths.
}
\description{
This function performs a chunked minimap2 alignment of two single-sequence FASTA files, generates a PAF file, and creates a dotplot PDF. It utilizes helper functions for sequence chunking, minimap2 alignment, PAF recombination, and dotplot generation.
}
\author{
Wolfram Höps
}
