% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{make_output_folder_structure}
\alias{make_output_folder_structure}
\title{make_output_folder_structure}
\usage{
make_output_folder_structure(sequence_name_output)
}
\description{
Create empty output folders that can be filled with data later.
}
