% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_remove_zero_pads.R
\name{matrix_remove_zero_pads}
\alias{matrix_remove_zero_pads}
\title{Quick and dirty way of getting rid of padding zero columns or rows.}
\usage{
matrix_remove_zero_pads(bitl_f)
}
\description{
Quick and dirty way of getting rid of padding zero columns or rows.
}
\author{
Wolfram Häps
}
