% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_paf.R
\name{merge_paf_entries_intraloop}
\alias{merge_paf_entries_intraloop}
\title{Tiny undocumented helperfunction.}
\usage{
merge_paf_entries_intraloop(inpaf)
}
\description{
Tiny undocumented helperfunction.
}
\author{
Wolfram Höps
}
