% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_paf.R
\name{merge_rows}
\alias{merge_rows}
\title{Helperfunction (1/1) of compress_paf_fnct}
\usage{
merge_rows(paffile, nl1, nl2)
}
\arguments{
\item{paffile}{\link{data frame} loaded paffile with colnames}

\item{nl1}{\link{character/link} line number of first-of-pair}

\item{nl2}{\link{character/link} line number of second-of-pair}
}
\value{
paffile with one row less (bc a pair has been merged)
}
\description{
Melt two alignment entries into one.
}
\author{
Wolfram Höps
}
