% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nahrwhals_main.R
\name{nahrwhals}
\alias{nahrwhals}
\title{Nahrwhals: A function for analyzing genomic alignments and visualizing results}
\usage{
nahrwhals(
  genome_x_fa,
  genome_y_fa,
  seqname_x,
  start_x,
  end_x,
  genome_y_fa_mmi = "default",
  anntrack = FALSE,
  logfile = "res/res.tsv",
  samplename_x = "Fasta_x",
  samplename_y = "Fasta_y",
  compare_full_fastas = FALSE,
  plot_only = FALSE,
  self_plots = TRUE,
  plot_xy_segmented = TRUE,
  eval_th = 98,
  depth = 3,
  chunklen = "default",
  minlen = "default",
  compression = "default",
  max_size_col_plus_rows = 250,
  max_n_alns = 150,
  use_paf_library = FALSE,
  conversionpaf_link = FALSE,
  xpad = 1,
  plot_minlen = 350,
  maxlen_refine = 1e+10,
  n_tests = 10,
  n_max_testchunks = 5,
  baseline_log_minsize_min = 8,
  baseline_log_minsize_max = FALSE,
  discovery_exact = FALSE,
  hltrack = FALSE,
  hllink = FALSE,
  aln_pad_factor = 1,
  debug = FALSE,
  clean_after_yourself = FALSE,
  testrun_std = FALSE,
  testrun_fullfa = FALSE,
  minimap2_bin = "default",
  bedtools_bin = "default"
)
}
\arguments{
\item{genome_x_fa}{Path to reference fasta file (default: "testdata/assemblies/hg38_partial.fa").}

\item{genome_y_fa}{Path to query fasta file (default: "testdata/assemblies/assembly_partial.fa").}

\item{seqname_x}{Sequence name for the reference fasta file (default: "chr1_partial").}

\item{start_x}{Start position for the reference sequence (default: 1700000).}

\item{end_x}{End position for the reference sequence (default: 3300000).}

\item{genome_y_fa_mmi}{Path to the .mmi index of the query fasta file or "default" (default: "default").}

\item{anntrack}{Logical, whether to use annotation track (default: FALSE).}

\item{logfile}{Path to output log file (default: "res/unittest.tsv").}

\item{samplename_x}{Sample name for the reference fasta file (default: "Fasta_x").}

\item{samplename_y}{Sample name for the query fasta file (default: "Fasta_y").}

\item{compare_full_fastas}{Logical, whether to compare full fasta files (default: FALSE).}

\item{plot_only}{Logical, whether to perform plotting only (default: FALSE).}

\item{self_plots}{Logical, whether to generate self plots (default: TRUE).}

\item{plot_xy_segmented}{Logical, whether to generate segmented XY plots (default: TRUE).}

\item{eval_th}{Evaluation threshold (default: 98).}

\item{depth}{Depth of the analysis (default: 3).}

\item{chunklen}{Chunk length or "default" (default: "default").}

\item{minlen}{Minimum length or "default" (default: "default").}

\item{compression}{Compression or "default" (default: "default").}

\item{max_size_col_plus_rows}{Maximum sum of rows and columns for alignment plots (default: 250).}

\item{max_n_alns}{Maximum number of alignments to process (default: 150).}

\item{use_paf_library}{Logical, whether to use the PAF library (default: FALSE).}

\item{conversionpaf_link}{Logical, whether to use the ConversionPAF link (default: FALSE).}

\item{xpad}{Padding for the X-axis (default: 1).}

\item{plot_minlen}{Minimum length for plotting (default: 350).}

\item{maxlen_refine}{Maximum length to refine (default: 1e10).}

\item{n_tests}{Number of tests to perform (default: 10).}

\item{n_max_testchunks}{Maximum number of test chunks (default: 5).}

\item{baseline_log_minsize_min}{Minimum log size for the baseline (default: 8).}

\item{baseline_log_minsize_max}{Maximum log size for the baseline or "default" (default: FALSE).}

\item{discovery_exact}{Logical, whether to use exact discovery mode (default: FALSE).}

\item{hltrack}{Logical, whether to use the high-level track (default: FALSE).}

\item{hllink}{Logical, whether to use the high-level link (default: FALSE).}

\item{aln_pad_factor}{Alignment padding factor (default: 1.0).}

\item{debug}{Logical, whether to enable debug mode (default: FALSE).}

\item{clean_after_yourself}{Logical, whether to delete all non-plot files, such as fastas (default: FALSE)}

\item{testrun_std}{Logical, whether to perform a test/demo run. Mut. Exclusive with testrun_fullfa. (default: FALSE)}

\item{testrun_fullfa}{Logical, whether to perform a test/demo run of full fastas. Mut. Exclusive with testrun_std (default: FALSE).}

\item{minimap2_bin}{Path to minimap2 binary or "default" (default: "default").}

\item{bedtools_bin}{Path to bedtools binary or "default" (default: "default").}
}
\value{
No explicit return value; function generates output files and visualizations.
}
\description{
This function performs sequence alignment between two genomic sequences (fasta files) and
generates visualizations of the alignment results.
}
