% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paf_dotplot.R
\name{pafdotplot_make}
\alias{pafdotplot_make}
\title{Wrapperfunction for turning a paf (and an SD highlight annotation) into a dotplot pdf.}
\usage{
pafdotplot_make(
  inpaf_link,
  outplot_link,
  min_align = 11,
  min_query_aln = 11,
  keep_ref = 10000,
  similarity = T,
  h_lines = T,
  interactive = F,
  plot_size = 10,
  on_target = T,
  v = F,
  hllink = F,
  hltype = F,
  hlstart = NULL,
  hlend = NULL,
  save = T,
  minsdlen = 5000,
  anntrack = F,
  x_seqname = NULL,
  x_start = NULL,
  x_end = NULL,
  hltrack = NULL
)
}
\arguments{
\item{inpaf_link}{\link{character/link} a link to the paf to be plotted}

\item{outplot_link}{\link{character/link} a link to where the output pdf will be saved}

\item{min_align}{not sure. set to 11, keep it there.}

\item{min_query_aln}{not sure. set to 11, keep it there.}

\item{keep_ref}{number of sequences to keep or something. Dont touch.}

\item{similarity}{\link{T/F} indicate sequence similarity by color code (default: T)}

\item{h_lines}{\link{T/F} plot horizontal lines (default: T)}

\item{interactive}{\link{T/F} unsure. Dotplotly? Untested, keep at F.}

\item{plot_size}{\link{numeric} plot size in inch.}

\item{on_target}{\link{T/F} dont remember. default: T}

\item{v}{\link{T/F} unsure. default: F}

\item{hllink}{\link{character/link} link to an SD annotation file (bed, tsv or paf) to include as
highlights in the plot.}

\item{hltype}{\link{character} filetype of hllink. Can be 'NULL', 'bed', 'tsv', 'paf'.}
}
\value{
Nothing, but writes a pdf with a dotplot.
}
\description{
T
}
\author{
Wolfram Höps
}
