% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_ggplot.R
\name{plot_matrix_ggplot_named}
\alias{plot_matrix_ggplot_named}
\title{plot_matrix_ggplot_named}
\usage{
plot_matrix_ggplot_named(data_frame_xyz, colnames_f, rownames_f)
}
\arguments{
\item{data_frame_xyz}{a dataframe with three coordinate columns (x,y,z)}
}
\description{
Make a plot of a dataframe matrix.
}
\author{
Wolfram Hoeps
}
