% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_wrapper.R
\name{produce_pairwise_alignments_minimap2}
\alias{produce_pairwise_alignments_minimap2}
\title{Produce up to three minimap2 based alignments.}
\usage{
produce_pairwise_alignments_minimap2(params, outlinks, chr_start_end_pad)
}
\description{
Produce up to three minimap2 based alignments.
}
\author{
Wolfram Höps
}
