% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{reduce_depth_if_needed}
\alias{reduce_depth_if_needed}
\title{reduce_depth_if_needed}
\usage{
reduce_depth_if_needed(bitlocus, increase_only, maxdepth)
}
\description{
Called at the beginning of a tree search. If the dotplot is unreasonably large,
we pragmatically reduce the search depth here to prevent overly long runtimes.
}
