% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{save_plot_custom}
\alias{save_plot_custom}
\title{save_plot_custom}
\usage{
save_plot_custom(
  inplot,
  filename,
  device,
  width = 20,
  height = 20,
  units = "cm"
)
}
\description{
Quick helperfunction for saving a ggplot plot.
}
