% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_logfile.R
\name{save_to_logfile}
\alias{save_to_logfile}
\title{I apologize for the existence of this terrible function.}
\usage{
save_to_logfile(log, res, logfile, params, alt_x = F)
}
\description{
I apologize for the existence of this terrible function.
}
