% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2.R
\name{solve_mutation}
\alias{solve_mutation}
\title{Alternative wrapper for calling dfs several times.
THis function is called as a standard.
TODO: description}
\usage{
solve_mutation(bitlocus, maxdepth_input, earlystop = Inf)
}
\description{
Alternative wrapper for calling dfs several times.
THis function is called as a standard.
TODO: description
}
