% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mutation_space_v2_helpers.R
\name{transform_res_new_to_old}
\alias{transform_res_new_to_old}
\title{transform_res_new_to_old}
\usage{
transform_res_new_to_old(res_df_f)
}
\arguments{
\item{res_df_f}{The results of the tree search in the new format.}
}
\value{
The results of the tree search in the old format.
}
\description{
This function transforms between different ways of representing the results of a tree search.
}
