% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_aln_and_analyse.R
\name{wrapper_condense_paf}
\alias{wrapper_condense_paf}
\title{wrapper_condense_paf}
\usage{
wrapper_condense_paf(params, outlinks)
}
\description{
Wrapper around a wrapper ... consider to simplify / remove.
}
