% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_bitlocus.R
\name{wrapper_paf_to_bitlocus}
\alias{wrapper_paf_to_bitlocus}
\title{wrapper_paf_to_bitlocus}
\usage{
wrapper_paf_to_bitlocus(
  inpaf,
  realplot = T,
  bitlocusplot = T,
  saveplot = F,
  minlen = 1000,
  compression = 1000
)
}
\description{
wrapper_paf_to_bitlocus
}
\author{
Wolfram Hoeps
}
